# swagger.yaml
...
openapi: 3.0.0
paths:
...
  /api/bookings:
    get:
      description: Optional filters are combined. Returns ascending date/start order,
        including existing bookings for inactive vehicles.
      operationId: BookingsController_list
      parameters:
        - name: date
          required: false
          in: query
          description: Filter by local calendar date. Historical dates are allowed when
            listing bookings.
          schema:
            format: date
            pattern: ^\d{4}-\d{2}-\d{2}$
            example: 2026-10-10
            type: string
        - name: vehicleId
          required: false
          in: query
          description: Filter by an existing vehicle, including inactive vehicles.
          schema:
            pattern: ^[a-zA-Z0-9_-]{1,80}$
            example: vehicle_1
            type: string
        - name: type
          required: false
          in: query
          schema:
            $ref: "#/components/schemas/BookingType"
      responses:
        "200":
          description: Matching operations in date/start order.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingResponseDto"
              examples:
                schedule:
                  summary: A matching scheduled operation
                  value:
                    - title: Interior cleaning
                      description: Prepare for the next handoff.
                      type: CLEANING
                      date: 2026-10-10
                      startTime: 10:00
                      durationMinutes: 60
                      vehicleId: vehicle_1
                      id: booking_1
                      createdAt: 2026-10-09T08:00:00.000Z
                      updatedAt: 2026-10-09T08:00:00.000Z
                empty:
                  summary: No operations match the filters
                  value: []
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiErrorResponseDto"
                  - type: object
                    properties:
                      code:
                        type: string
                        enum:
                          - VALIDATION_ERROR
              examples:
                VALIDATION_ERROR:
                  value:
                    code: VALIDATION_ERROR
                    message: Invalid request fields. Check required values and supported formats.
        "404":
          description: VEHICLE_NOT_FOUND
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiErrorResponseDto"
                  - type: object
                    properties:
                      code:
                        type: string
                        enum:
                          - VEHICLE_NOT_FOUND
              examples:
                VEHICLE_NOT_FOUND:
                  value:
                    code: VEHICLE_NOT_FOUND
                    message: Vehicle not found.
      summary: List scheduled operations
      tags:
        - Bookings
    post:
      description: Prevents overlapping operations for the same vehicle and local
        date. Adjacent windows and simultaneous bookings on different vehicles
        are allowed. Unknown fields and null values are rejected.
      operationId: BookingsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingDto"
            examples:
              cleaning:
                summary: Schedule a 60-minute cleaning
                value:
                  title: Interior cleaning
                  description: Prepare for the next handoff.
                  type: CLEANING
                  date: 2026-10-10
                  startTime: 10:00
                  durationMinutes: 60
                  vehicleId: vehicle_1
              inspection:
                summary: Schedule a 30-minute inspection
                value:
                  title: Return inspection
                  description: Prepare for the next handoff.
                  type: INSPECTION
                  date: 2026-10-10
                  startTime: 10:00
                  durationMinutes: 30
                  vehicleId: vehicle_1
      responses:
        "201":
          description: Operation created with a generated ID and timestamps.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponseDto"
              example:
                title: Interior cleaning
                description: Prepare for the next handoff.
                type: CLEANING
                date: 2026-10-10
                startTime: 10:00
                durationMinutes: 60
                vehicleId: vehicle_1
                id: d6d6e4cd-7f42-4b8c-8a85-76fc7154315a
                createdAt: 2026-10-09T08:00:00.000Z
                updatedAt: 2026-10-09T08:00:00.000Z
        "400":
          description: VALIDATION_ERROR, PAST_DATE, INVALID_SLOT, INVALID_DURATION,
            OUTSIDE_WORKING_HOURS
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiErrorResponseDto"
                  - type: object
                    properties:
                      code:
                        type: string
                        enum:
                          - VALIDATION_ERROR
                          - PAST_DATE
                          - INVALID_SLOT
                          - INVALID_DURATION
                          - OUTSIDE_WORKING_HOURS
              examples:
                VALIDATION_ERROR:
                  value:
                    code: VALIDATION_ERROR
                    message: Invalid request fields. Check required values and supported formats.
                PAST_DATE:
                  value:
                    code: PAST_DATE
                    message: Bookings cannot be scheduled in the past.
                INVALID_SLOT:
                  value:
                    code: INVALID_SLOT
                    message: Choose a start time on the 30-minute grid.
                INVALID_DURATION:
                  value:
                    code: INVALID_DURATION
                    message: Duration must be 30–240 minutes in 30-minute increments.
                OUTSIDE_WORKING_HOURS:
                  value:
                    code: OUTSIDE_WORKING_HOURS
                    message: Operations must fit within 09:00–17:00.
        "404":
          description: VEHICLE_NOT_FOUND
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiErrorResponseDto"
                  - type: object
                    properties:
                      code:
                        type: string
                        enum:
                          - VEHICLE_NOT_FOUND
              examples:
                VEHICLE_NOT_FOUND:
                  value:
                    code: VEHICLE_NOT_FOUND
                    message: Vehicle not found.
        "409":
          description: VEHICLE_INACTIVE, BOOKING_CONFLICT
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                      - $ref: "#/components/schemas/ApiErrorResponseDto"
                      - type: object
                        properties:
                          code:
                            type: string
                            enum:
                              - VEHICLE_INACTIVE
                  - $ref: "#/components/schemas/BookingConflictResponseDto"
              examples:
                VEHICLE_INACTIVE:
                  value:
                    code: VEHICLE_INACTIVE
                    message: Bookings cannot be scheduled for an inactive vehicle.
                BOOKING_CONFLICT:
                  value:
                    code: BOOKING_CONFLICT
                    message: This vehicle is unavailable for the requested window.
                    suggestions:
                      - startTime: 11:00
                        endTime: 12:00
                      - startTime: 11:30
                        endTime: 12:30
                      - startTime: 12:00
                        endTime: 13:00
      summary: Schedule an operation
      tags:
        - Bookings
  /api/bookings/{id}:
    get:
      operationId: BookingsController_get
      parameters:
        - name: id
          required: true
          in: path
          schema:
            pattern: ^[a-zA-Z0-9_-]{1,80}$
            example: booking_1
            type: string
      responses:
        "200":
          description: The requested operation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponseDto"
              example:
                title: Interior cleaning
                description: Prepare for the next handoff.
                type: CLEANING
                date: 2026-10-10
                startTime: 10:00
                durationMinutes: 60
                vehicleId: vehicle_1
                id: booking_1
                createdAt: 2026-10-09T08:00:00.000Z
                updatedAt: 2026-10-09T08:00:00.000Z
        "404":
          description: BOOKING_NOT_FOUND
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiErrorResponseDto"
                  - type: object
                    properties:
                      code:
                        type: string
                        enum:
                          - BOOKING_NOT_FOUND
              examples:
                BOOKING_NOT_FOUND:
                  value:
                    code: BOOKING_NOT_FOUND
                    message: Booking not found.
      summary: Get one operation
      tags:
        - Bookings
    patch:
      description: All fields are optional. Submitted fields are merged with the
        stored booking, then the complete candidate is revalidated. The current
        booking is excluded from overlap detection. Omitted fields are
        preserved; explicit null values and unknown fields are rejected.
      operationId: BookingsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            pattern: ^[a-zA-Z0-9_-]{1,80}$
            example: booking_1
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingDto"
            examples:
              reschedule:
                summary: Move the existing operation; preserve other fields
                value:
                  startTime: 11:00
              rename:
                summary: Change only the title
                value:
                  title: Afternoon cleaning
              clearNotes:
                summary: Clear optional notes
                value:
                  description: ""
      responses:
        "200":
          description: The complete operation after merging and validating the patch.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponseDto"
              example:
                title: Interior cleaning
                description: Prepare for the next handoff.
                type: CLEANING
                date: 2026-10-10
                startTime: 11:00
                durationMinutes: 60
                vehicleId: vehicle_1
                id: booking_1
                createdAt: 2026-10-09T08:00:00.000Z
                updatedAt: 2026-10-09T08:30:00.000Z
        "400":
          description: VALIDATION_ERROR, PAST_DATE, INVALID_SLOT, INVALID_DURATION,
            OUTSIDE_WORKING_HOURS
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiErrorResponseDto"
                  - type: object
                    properties:
                      code:
                        type: string
                        enum:
                          - VALIDATION_ERROR
                          - PAST_DATE
                          - INVALID_SLOT
                          - INVALID_DURATION
                          - OUTSIDE_WORKING_HOURS
              examples:
                VALIDATION_ERROR:
                  value:
                    code: VALIDATION_ERROR
                    message: Invalid request fields. Check required values and supported formats.
                PAST_DATE:
                  value:
                    code: PAST_DATE
                    message: Bookings cannot be scheduled in the past.
                INVALID_SLOT:
                  value:
                    code: INVALID_SLOT
                    message: Choose a start time on the 30-minute grid.
                INVALID_DURATION:
                  value:
                    code: INVALID_DURATION
                    message: Duration must be 30–240 minutes in 30-minute increments.
                OUTSIDE_WORKING_HOURS:
                  value:
                    code: OUTSIDE_WORKING_HOURS
                    message: Operations must fit within 09:00–17:00.
        "404":
          description: VEHICLE_NOT_FOUND, BOOKING_NOT_FOUND
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiErrorResponseDto"
                  - type: object
                    properties:
                      code:
                        type: string
                        enum:
                          - VEHICLE_NOT_FOUND
                          - BOOKING_NOT_FOUND
              examples:
                VEHICLE_NOT_FOUND:
                  value:
                    code: VEHICLE_NOT_FOUND
                    message: Vehicle not found.
                BOOKING_NOT_FOUND:
                  value:
                    code: BOOKING_NOT_FOUND
                    message: Booking not found.
        "409":
          description: VEHICLE_INACTIVE, BOOKING_CONFLICT
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                      - $ref: "#/components/schemas/ApiErrorResponseDto"
                      - type: object
                        properties:
                          code:
                            type: string
                            enum:
                              - VEHICLE_INACTIVE
                  - $ref: "#/components/schemas/BookingConflictResponseDto"
              examples:
                VEHICLE_INACTIVE:
                  value:
                    code: VEHICLE_INACTIVE
                    message: Bookings cannot be scheduled for an inactive vehicle.
                BOOKING_CONFLICT:
                  value:
                    code: BOOKING_CONFLICT
                    message: This vehicle is unavailable for the requested window.
                    suggestions:
                      - startTime: 11:00
                        endTime: 12:00
                      - startTime: 11:30
                        endTime: 12:30
                      - startTime: 12:00
                        endTime: 13:00
      summary: Edit or reschedule an operation
      tags:
        - Bookings
    delete:
      description: Hard deletion. Successful responses have no body.
      operationId: BookingsController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            pattern: ^[a-zA-Z0-9_-]{1,80}$
            example: booking_1
            type: string
      responses:
        "204":
          description: Operation deleted; empty response body.
        "404":
          description: BOOKING_NOT_FOUND
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiErrorResponseDto"
                  - type: object
                    properties:
                      code:
                        type: string
                        enum:
                          - BOOKING_NOT_FOUND
              examples:
                BOOKING_NOT_FOUND:
                  value:
                    code: BOOKING_NOT_FOUND
                    message: Booking not found.
      summary: Delete an operation
      tags:
        - Bookings
info:
  title: FleetSlot API
  description: Fleet turnaround and vehicle operations scheduling. In-memory
    fictional data resets on backend restart. Dates use the operator/server
    local calendar; replace example dates with today or a future date before
    scheduling. No authentication is required for this demo.
  version: 1.0.0
  contact: {}
...
servers:
  - url: /
    description: Current host (API or frontend proxy)
components:
  schemas:
...
    ApiErrorResponseDto:
      type: object
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - PAST_DATE
            - INVALID_SLOT
            - INVALID_DURATION
            - OUTSIDE_WORKING_HOURS
            - VEHICLE_NOT_FOUND
            - BOOKING_NOT_FOUND
            - VEHICLE_INACTIVE
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid request fields. Check required values and supported formats.
      required:
        - code
        - message
    TimeWindowDto:
      type: object
      properties:
        startTime:
          type: string
          pattern: ^([01]\d|2[0-3]):[0-5]\d$
          example: 10:00
          description: Local HH:mm time, aligned to 30-minute starts. The whole operation
            must fit within 09:00–17:00.
        endTime:
          type: string
          pattern: ^([01]\d|2[0-3]):[0-5]\d$
          example: 11:00
          description: Derived local end time. Never submitted or persisted as a booking
            field.
      required:
        - startTime
        - endTime
    BookingConflictResponseDto:
      type: object
      properties:
        code:
          type: string
          enum:
            - BOOKING_CONFLICT
        message:
          type: string
          example: This vehicle is unavailable for the requested window.
        suggestions:
          maxItems: 3
          description: Same vehicle, date, and duration. Later starts ascending, then
            earlier starts descending. May be empty when the day is fully
            booked.
          type: array
          items:
            $ref: "#/components/schemas/TimeWindowDto"
      required:
        - code
        - message
        - suggestions
...
    BookingType:
      type: string
      enum:
        - PICKUP_HANDOFF
        - RETURN_HANDOFF
        - INSPECTION
        - CLEANING
        - MAINTENANCE
    CreateBookingDto:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
          example: Interior cleaning
          description: Trimmed before validation; cannot be blank.
        description:
          type: string
          maxLength: 1000
          example: Prepare for the next handoff.
          description: Optional notes. On update, omit to preserve notes or send an empty
            string to clear them; null is not accepted.
        type:
          example: CLEANING
          allOf:
            - $ref: "#/components/schemas/BookingType"
        date:
          type: string
          format: date
          pattern: ^\d{4}-\d{2}-\d{2}$
          example: 2026-10-10
          description: Plain local calendar date (YYYY-MM-DD). Scheduling and availability
            require today or later; replace example dates before trying
            requests.
        startTime:
          type: string
          pattern: ^([01]\d|2[0-3]):[0-5]\d$
          example: 10:00
          description: Local HH:mm time, aligned to 30-minute starts. The whole operation
            must fit within 09:00–17:00.
        durationMinutes:
          type: integer
          minimum: 30
          maximum: 240
          multipleOf: 30
          example: 60
          description: Operation duration in minutes.
        vehicleId:
          type: string
          pattern: ^[a-zA-Z0-9_-]{1,80}$
          example: vehicle_1
          description: Vehicle ID. Creating or updating a booking requires an active
            vehicle.
      required:
        - title
        - type
        - date
        - startTime
        - durationMinutes
        - vehicleId
      additionalProperties: false
    UpdateBookingDto:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
          example: Interior cleaning
          description: Trimmed before validation; cannot be blank.
        description:
          type: string
          maxLength: 1000
          example: Prepare for the next handoff.
          description: Optional notes. On update, omit to preserve notes or send an empty
            string to clear them; null is not accepted.
        type:
          example: CLEANING
          allOf:
            - $ref: "#/components/schemas/BookingType"
        date:
          type: string
          format: date
          pattern: ^\d{4}-\d{2}-\d{2}$
          example: 2026-10-10
          description: Plain local calendar date (YYYY-MM-DD). Scheduling and availability
            require today or later; replace example dates before trying
            requests.
        startTime:
          type: string
          pattern: ^([01]\d|2[0-3]):[0-5]\d$
          example: 10:00
          description: Local HH:mm time, aligned to 30-minute starts. The whole operation
            must fit within 09:00–17:00.
        durationMinutes:
          type: integer
          minimum: 30
          maximum: 240
          multipleOf: 30
          example: 60
          description: Operation duration in minutes.
        vehicleId:
          type: string
          pattern: ^[a-zA-Z0-9_-]{1,80}$
          example: vehicle_1
          description: Vehicle ID. Creating or updating a booking requires an active
            vehicle.
      additionalProperties: false
    BookingResponseDto:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
          example: Interior cleaning
          description: Trimmed before validation; cannot be blank.
        description:
          type: string
          maxLength: 1000
          example: Prepare for the next handoff.
          description: Optional notes. On update, omit to preserve notes or send an empty
            string to clear them; null is not accepted.
        type:
          example: CLEANING
          allOf:
            - $ref: "#/components/schemas/BookingType"
        date:
          type: string
          format: date
          pattern: ^\d{4}-\d{2}-\d{2}$
          example: 2026-10-10
          description: Plain local calendar date (YYYY-MM-DD). Scheduling and availability
            require today or later; replace example dates before trying
            requests.
        startTime:
          type: string
          pattern: ^([01]\d|2[0-3]):[0-5]\d$
          example: 10:00
          description: Local HH:mm time, aligned to 30-minute starts. The whole operation
            must fit within 09:00–17:00.
        durationMinutes:
          type: integer
          minimum: 30
          maximum: 240
          multipleOf: 30
          example: 60
          description: Operation duration in minutes.
        vehicleId:
          type: string
          pattern: ^[a-zA-Z0-9_-]{1,80}$
          example: vehicle_1
          description: Vehicle ID. Creating or updating a booking requires an active
            vehicle.
        id:
          type: string
          pattern: ^[a-zA-Z0-9_-]{1,80}$
          example: booking_1
          readOnly: true
          description: Seed IDs are stable strings; new bookings receive UUIDs.
        createdAt:
          type: string
          format: date-time
          example: 2026-10-09T08:00:00.000Z
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          example: 2026-10-09T08:00:00.000Z
          readOnly: true
      required:
        - title
        - type
        - date
        - startTime
        - durationMinutes
        - vehicleId
        - id
        - createdAt
        - updatedAt